% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chictr_read_results.R
\name{chictr_read_results}
\alias{chictr_read_results}
\title{Search ChiCTR and read the results table}
\usage{
chictr_read_results(URL)
}
\arguments{
\item{URL}{The URL given after searching on ClinicalTrials.gov.}
}
\value{
A dataframe of the ChiCTR results table.
}
\description{
First, navigate to www.chictr.org.cn and create the search term of interest.
    Then, copy-paste the resulting URL into this function. The function will
    download the registration number, public title, type, registration time
    and a linke to the full article.
}
\examples{
\dontrun{
# Copy-paste the URL from the ChiCTR.org.cn website
URL <- "http://www.chictr.org.cn/searchprojen.aspx?title=covid"
# Download results table
results_table <- chictr_read_results(URL)
}
}
