% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chictr_read_trial.R
\name{chictr_read_trial}
\alias{chictr_read_trial}
\title{Read fields of interest from a specific trial}
\usage{
chictr_read_trial(trial_url)
}
\arguments{
\item{trial_url}{The URL of a specific trial as a string}
}
\value{
A tibble of the trial's most commonly used fields.
}
\description{
Takes the URL of a specific trial, downloads the XML file, reads the fields
    that are most commonly of interest and returns a tibble.
}
\examples{
\dontrun{
# Copy-paste the trial URL from the ChiCTR.org.cn website
trial_url <- "http://www.chictr.org.cn/showprojen.aspx?proj=55654"
# Extract the fields of interest
trial <- chictr_read_trial(trial_url)
}
}
