% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chictr_write_trial.R
\name{chictr_write_trial}
\alias{chictr_write_trial}
\title{Download a trial as an XML}
\usage{
chictr_write_trial(trial_url, path = NULL, ...)
}
\arguments{
\item{trial_url}{The URL of a specific trial as a string.}

\item{path}{The path to the directory in which the file should be saved. Use
NULL to save in the current directory (default).}

\item{...}{Additional arguments to be passed to `xml2::write_xml`.}
}
\value{
The study record as an XML file saved at the designed path.
}
\description{
Takes the URL of a specific trial, and downloads the available data as an
    XML.
}
\examples{
\dontrun{
# Copy-paste the trial URL from the ChiCTR.org.cn website
trial_url <- "http://www.chictr.org.cn/showprojen.aspx?proj=55654"
# Extract the fields of interest
chictr_write_trial(trial_url)
}
}
