% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchictr.R
\docType{package}
\name{clinicaltrialr-package}
\alias{clinicaltrialr}
\alias{clinicaltrialr-package}
\title{rchictr: Import study records from chictr.org.cn}
\description{
It retrieves the full study records of the desired studies from
chictr.org.cn and downloads a dataset of the most commonly used fields.
}
\author{
\strong{Maintainer}: Stylianos Serghiou \email{stelios.serghiou@gmail.com}

}
