% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chictr_read_results.R
\name{.extract_tables}
\alias{.extract_tables}
\title{Extract the results table from the ChiCTR results page}
\usage{
.extract_tables(index, URL, n_pages)
}
\arguments{
\item{index}{The page number to extract from as an integer or string.}

\item{URL}{The URL of the results page after searching on chictr.org.cn.}
}
\value{
A dataframe of the ChiCTR results table.
}
\description{
Extracts the results table from each successive page of search results.
}
\examples{
\dontrun{
# Copy-paste the URL from the ChiCTR.org.cn website
URL <- "http://www.chictr.org.cn/searchprojen.aspx?title=covid"
# Download results table
results_table <- chictr_read_results(URL)
}
}
