% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chictr_read_trial.R
\name{.get_text}
\alias{.get_text}
\title{Get the desired text from the xml_document}
\usage{
.get_text(xml_document, xpath)
}
\arguments{
\item{xml_document}{The ChiCTR trial as an XML document}

\item{xpath}{The XPath as a character, e.g. "Triall/main/trial_id"}
}
\value{
The desired text as a character; if not found, then `character()`
}
\description{
Returns the text desired according to XPath.
}
